Refactor quantifiers
